support npm@3
The issue arises due to npm@3's more agressive tendency towards "flat
dependencies". This leaves the reporter dependency in a location that
jenkins-mocha isn't expecting, when jenkins-mocha is installed into a
parent project. The solution here is to use require.resolve in order to
figure out where the reporter dependency ought to be found.